Add files via upload
The code used for creating the 6 tables in pgAdmin. While host_review_stats has host_id and host_location_review_stats has host_location_id, both of which are in listing_info, in listing_info they are duplicate values and therefore cannot be used as keys. In listing_info the only unique value is listing_id, which I cannot call upon in host_review_stats or host_location_review_stats because in many cases there are several listing_ids for each host or each location. Hence, the tables are related, but not tied by foreign keys.