Add files via upload
The schema is up to date with what was loaded into the dataframe and the ERB code (in hastags at the bottom) in reading_etl_csvs is also up to date. The elt_project.ipynb reads the csvs, divides the data into the desired tables, cleans the data, and uses SQL Alchemy to connect and load the new tables into the database in pgAdmin.