Add files via upload
This .png is not an exact representation of the code I did, but rather a guide as to how columns in the table connect to one another. This is because averaged the review scores for both hosts and locations, meaning that while there is only one host_id and host_location, the same values occur multiple times in the listings_info table. Since they are not unique values in the listings_info table, I did not tie to them using foreign keys. However, this diagram can be useful show how the columns are related to one another: just do not take it as a literal picture of how I wrote my code.